feat(notifications): created a Helper to place functions in a single location
Created a helper file, this makes it easier to make changes to the function that multiple different
agents use in the same fashion.

BREAKING CHANGE: Adding extra content to notifications

closes #2104
