+++ server/lib/notifications/agents/email.ts
@@ -1,12 +1,15 @@
 import { IssueType, IssueTypeName } from '@server/constants/issue';
 import { MediaType } from '@server/constants/media';
-import { MediaServerType } from '@server/constants/server';
 import { getRepository } from '@server/datasource';
 import { User } from '@server/entity/User';
 import PreparedEmail from '@server/lib/email';
 import type { NotificationAgentEmail } from '@server/lib/settings';
 import { NotificationAgentKey, getSettings } from '@server/lib/settings';
 import logger from '@server/logger';
+import {
+  getAvailableMediaServerName,
+  getAvailableMediaServerUrl,
+} from '@server/utils/mediaServerHelper';
 import type { EmailOptions } from 'email-templates';
 import path from 'path';
 import validator from 'validator';
@@ -52,30 +55,8 @@ class EmailAgent
     const settings = getSettings();
     const { applicationUrl, applicationTitle, mediaServerType } = settings.main;
     const { embedPoster } = settings.notifications.agents.email;
-
-    function getAvailableMediaServerName() {
-      if (mediaServerType === MediaServerType.EMBY) {
-        return 'Emby';
-      }
-
-      if (mediaServerType === MediaServerType.PLEX) {
-        return 'Plex';
-      }
-
-      return 'Jellyfin';
-    }
-
-    const mediaServerName = getAvailableMediaServerName();
-
-    function getAvailableMediaServerUrl(): string | undefined {
-      const wants4k = payload.request?.is4k;
-      const url4k = (payload.media as any)?.mediaUrl4k as string | undefined;
-      const url = (payload.media as any)?.mediaUrl as string | undefined;
-
-      return (wants4k ? (url4k ?? url) : (url ?? url4k)) || undefined;
-    }
-
-    const mediaServerUrl = getAvailableMediaServerUrl();
+    const mediaServerName = getAvailableMediaServerName(mediaServerType);
+    const mediaServerUrl = getAvailableMediaServerUrl(payload);
 
     if (type === Notification.TEST_NOTIFICATION) {
       return {
